requirements.txt
# Core ML and AI libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
peft>=0.4.0
bitsandbytes>=0.41.0

# Model evaluation and metrics
evaluate>=0.4.0
rouge-score>=0.1.2
nltk>=3.8.1
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# API and web frameworks
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0
openai>=0.28.0
anthropic>=0.5.0

# Data processing and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0

# Jupyter and visualization
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Testing and development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Additional utilities
sentence-transformers>=2.2.0
sacrebleu>=2.3.0
bert-score>=0.3.13
